backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
coverage==7.13.0
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
bcrypt==4.0.1
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
passlib[bcrypt]==1.7.4
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-multipart==0.0.19
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
